refactor(web): re-organise folders and layouts
