+++ apps/web/app/(home)/components/header/userIconButton.tsx
@@ -3,7 +3,7 @@
 import React, { useState } from 'react'
 import { Button } from 'ui'
 import { UserIcon } from 'ui/icons'
-import IconButton from '../iconButton'
+import { IconButton } from '@/components'
 
 interface UserIconButtonProps extends React.HTMLAttributes<HTMLDivElement> {}
