+++ apps/web/app/(home)/components/index.ts
@@ -1,2 +1,3 @@
 export {default as CardList} from './cardList'
 export {default as Filter} from './filter'
+export {default as Header} from './header'
