+++ apps/web/app/(home)/layout.tsx
@@ -1,0 +1,16 @@
+import * as React from 'react'
+import { Header } from './components';
+
+export default function MainLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+
+  return (
+    <>
+      <Header/>
+      {children}
+    </>
+  );
+}
