+++ apps/web/app/components/index.ts
@@ -1,2 +1,2 @@
-export {default as Header} from './header'
 export {default as IconButton} from './iconButton'
+export {default as Themer} from './themer'
