+++ apps/web/app/components/themer.tsx
@@ -1,0 +1,14 @@
+'use client'
+
+import React, { useEffect } from 'react'
+
+export default function Themer() {
+  useEffect(() => {
+    if (window.matchMedia('(prefers-color-scheme: dark)').matches)
+      document.documentElement.classList.add('dark')
+    else
+      document.documentElement.classList.remove('dark')
+  },[])
+
+  return (<></>)
+}
